+++ .gitignore
@@ -2,5 +2,3 @@
 /node_modules
 package-lock.json
 shrinkwrap.yaml
-Output.txt
-xplevel.db
