--- commands/sort.js
@@ -1,36 +1,0 @@
-const Discord = require("discord.js");
-const client = new Discord.Client({
-  partials: ["MESSAGE", "CHANNEL", "REACTION"]
-});
-const moment = require("moment");
-const fs = require("fs");
-module.exports = {
-  name: "sort",
-  description: "Hopefully makes a graph of member count",
-  usage: "",
-  execute(message, args) {
-    message.reply("yes");
-    var joins=["2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-16","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-17","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-18","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-19","2018-10-20","2018-10-20","2018-10-20","2018-10-20","2018-10-20","2018-10-21","2018-10-21","2018-10-21","2018-10-21","2018-10-21","2018-10-21","2018-10-21","2018-10-21","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-22","2018-10-23","2018-10-23","2018-10-23","2018-10-23","2018-10-23","2018-10-23","2018-10-23","2018-10-24","2018-10-24","2018-10-24","2018-10-24","2018-10-24","2018-10-24","2018-10-24","2018-10-24","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-25","2018-10-26","2018-10-26","2018-10-26","2018-10-27","2018-10-27","2018-10-27","2018-10-27","2018-10-28","2018-10-28","2018-10-29","2018-10-29","2018-10-29","2018-10-29","2018-10-29","2018-10-29","2018-10-30","2018-10-30","2018-10-30","2018-10-30","2018-10-31","2018-11-1","2018-11-1","2018-11-1","2018-11-1","2018-11-1","2018-11-1","2018-11-10","2018-11-10","2018-11-10","2018-11-10","2018-11-10","2018-11-11","2018-11-11","2018-11-12","2018-11-12","2018-11-12","2018-11-12","2018-11-12","2018-11-12","2018-11-13","2018-11-13","2018-11-13","2018-11-13","2018-11-14","2018-11-14","2018-11-14","2018-11-15","2018-11-15","2018-11-15","2018-11-15","2018-11-15","2018-11-16","2018-11-16","2018-11-16","2018-11-16","2018-11-16","2018-11-16","2018-11-16","2018-11-16","2018-11-17","2018-11-17","2018-11-17","2018-11-17","2018-11-17","2018-11-17","2018-11-17","2018-11-18","2018-11-18","2018-11-18","2018-11-18","2018-11-18","2018-11-18","2018-11-18","2018-11-19","2018-11-19","2018-11-19","2018-11-19","2018-11-19","2018-11-19","2018-11-2","2018-11-2","2018-11-2","2018-11-2","2018-11-20","2018-11-20","2018-11-20","2018-11-20","2018-11-20","2018-11-20","2018-11-20","2018-11-21","2018-11-22","2018-11-22","2018-11-22","2018-11-23","2018-11-23","2018-11-23","2018-11-23","2018-11-24","2018-11-24","2018-11-24","2018-11-24","2018-11-25","2018-11-25","2018-11-25","2018-11-25","2018-11-26","2018-11-26","2018-11-26","2018-11-26","2018-11-27","2018-11-27","2018-11-27","2018-11-27","2018-11-27","2018-11-28","2018-11-28","2018-11-28","2018-11-28","2018-11-28","2018-11-28","2018-11-28","2018-11-29","2018-11-29","2018-11-29","2018-11-29","2018-11-29","2018-11-29","2018-11-29","2018-11-3","2018-11-3","2018-11-30","2018-11-30","2018-11-30","2018-11-30","2018-11-30","2018-11-30","2018-11-4","2018-11-4","2018-11-4","2018-11-4","2018-11-4","2018-11-5","2018-11-5","2018-11-6","2018-11-7","2018-11-7","2018-11-7","2018-11-7","2018-11-7","2018-11-8","2018-11-8","2018-11-8","2018-11-8","2018-11-9","2018-12-1","2018-12-1","2018-12-1","2018-12-1","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-10","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-11","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-12","2018-12-13","2018-12-13","2018-12-13","2018-12-13","2018-12-13","2018-12-13","2018-12-14","2018-12-14","2018-12-15","2018-12-15","2018-12-15","2018-12-15","2018-12-15","2018-12-15","2018-12-15","2018-12-15","2018-12-16","2018-12-16","2018-12-16","2018-12-16","2018-12-16","2018-12-16","2018-12-17","2018-12-17","2018-12-17","2018-12-18","2018-12-18","2018-12-18","2018-12-18","2018-12-18","2018-12-18","2018-12-19","2018-12-19","2018-12-19","2018-12-19","2018-12-19","2018-12-2","2018-12-2","2018-12-2","2018-12-2","2018-12-2","2018-12-2","2018-12-20","2018-12-20","2018-12-20","2018-12-20","2018-12-20","2018-12-20","2018-12-20","2018-12-20","2018-12-21","2018-12-21","2018-12-21","2018-12-21","2018-12-21","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-22","2018-12-23","2018-12-23","2018-12-23","2018-12-23","2018-12-24","2018-12-24","2018-12-24","2018-12-25","2018-12-25","2018-12-25","2018-12-25","2018-12-26","2018-12-26","2018-12-26","2018-12-26","2018-12-26","2018-12-26","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-27","2018-12-28","2018-12-28","2018-12-28","2018-12-28","2018-12-28","2018-12-29","2018-12-29","2018-12-29","2018-12-29","2018-12-29","2018-12-29","2018-12-29","2018-12-29","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-3","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-30","2018-12-31","2018-12-31","2018-12-31","2018-12-31","2018-12-31","2018-12-31","2018-12-4","2018-12-4","2018-12-4","2018-12-4","2018-12-5","2018-12-5","2018-12-5","2018-12-5","2018-12-6","2018-12-6","2018-12-6","2018-12-6","2018-12-6","2018-12-6","2018-12-6","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-7","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-8","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2018-12-9","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-1","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-10","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-11","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-12","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-13","2019-1-14","2019-1-14","2019-1-14","2019-1-14","2019-1-14","2019-1-14","2019-1-14","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-15","2019-1-16","2019-1-16","2019-1-16","2019-1-16","2019-1-16","2019-1-17","2019-1-17","2019-1-17","2019-1-17","2019-1-17","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-18","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-19","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-2","2019-1-20","2019-1-20","2019-1-20","2019-1-20","2019-1-20","2019-1-21","2019-1-21","2019-1-21","2019-1-21","2019-1-21","2019-1-21","2019-1-21","2019-1-21","2019-1-22","2019-1-22","2019-1-22","2019-1-22","2019-1-22","2019-1-22","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-23","2019-1-24","2019-1-24","2019-1-24","2019-1-24","2019-1-24","2019-1-24","2019-1-24","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-25","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-26","2019-1-27","2019-1-27","2019-1-27","2019-1-27","2019-1-27","2019-1-28","2019-1-28","2019-1-28","2019-1-28","2019-1-28","2019-1-28","2019-1-28","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-29","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-3","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-30","2019-1-31","2019-1-31","2019-1-31","2019-1-31","2019-1-31","2019-1-31","2019-1-4","2019-1-4","2019-1-4","2019-1-4","2019-1-4","2019-1-4","2019-1-4","2019-1-4","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-5","2019-1-6","2019-1-6","2019-1-6","2019-1-6","2019-1-6","2019-1-6","2019-1-6","2019-1-6","2019-1-7","2019-1-7","2019-1-7","2019-1-7","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-8","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-1-9","2019-10-1","2019-10-1","2019-10-10","2019-10-10","2019-10-10","2019-10-10","2019-10-11","2019-10-11","2019-10-11","2019-10-11","2019-10-12","2019-10-12","2019-10-12","2019-10-13","2019-10-13","2019-10-14","2019-10-14","2019-10-14","2019-10-14","2019-10-14","2019-10-15","2019-10-15","2019-10-15","2019-10-15","2019-10-15","2019-10-15","2019-10-17","2019-10-17","2019-10-17","2019-10-17","2019-10-18","2019-10-18","2019-10-19","2019-10-19","2019-10-19","2019-10-2","2019-10-20","2019-10-20","2019-10-20","2019-10-20","2019-10-20","2019-10-20","2019-10-20","2019-10-20","2019-10-21","2019-10-21","2019-10-21","2019-10-21","2019-10-21","2019-10-21","2019-10-22","2019-10-22","2019-10-22","2019-10-23","2019-10-23","2019-10-23","2019-10-23","2019-10-24","2019-10-24","2019-10-25","2019-10-25","2019-10-25","2019-10-25","2019-10-26","2019-10-26","2019-10-26","2019-10-26","2019-10-26","2019-10-27","2019-10-27","2019-10-28","2019-10-28","2019-10-28","2019-10-29","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-3","2019-10-30","2019-10-30","2019-10-30","2019-10-30","2019-10-31","2019-10-31","2019-10-4","2019-10-4","2019-10-5","2019-10-6","2019-10-6","2019-10-6","2019-10-6","2019-10-7","2019-10-7","2019-10-7","2019-10-7","2019-10-8","2019-10-8","2019-10-9","2019-10-9","2019-10-9","2019-10-9","2019-10-9","2019-10-9","2019-11-1","2019-11-1","2019-11-1","2019-11-10","2019-11-10","2019-11-10","2019-11-10","2019-11-10","2019-11-10","2019-11-11","2019-11-11","2019-11-11","2019-11-12","2019-11-12","2019-11-13","2019-11-13","2019-11-14","2019-11-14","2019-11-14","2019-11-14","2019-11-14","2019-11-14","2019-11-15","2019-11-15","2019-11-15","2019-11-15","2019-11-15","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-16","2019-11-17","2019-11-17","2019-11-18","2019-11-19","2019-11-2","2019-11-2","2019-11-2","2019-11-2","2019-11-2","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-20","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-21","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-22","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-23","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-24","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-25","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-26","2019-11-27","2019-11-27","2019-11-27","2019-11-27","2019-11-27","2019-11-27","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-28","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-29","2019-11-3","2019-11-3","2019-11-3","2019-11-3","2019-11-3","2019-11-30","2019-11-30","2019-11-30","2019-11-4","2019-11-4","2019-11-4","2019-11-5","2019-11-6","2019-11-6","2019-11-6","2019-11-6","2019-11-6","2019-11-6","2019-11-6","2019-11-7","2019-11-7","2019-11-7","2019-11-8","2019-11-8","2019-11-8","2019-11-8","2019-11-8","2019-11-8","2019-11-8","2019-11-8","2019-11-9","2019-11-9","2019-11-9","2019-11-9","2019-11-9","2019-11-9","2019-11-9","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-1","2019-12-10","2019-12-10","2019-12-10","2019-12-10","2019-12-10","2019-12-10","2019-12-10","2019-12-10","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-11","2019-12-12","2019-12-12","2019-12-12","2019-12-12","2019-12-12","2019-12-12","2019-12-12","2019-12-13","2019-12-13","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-14","2019-12-15","2019-12-15","2019-12-15","2019-12-15","2019-12-16","2019-12-16","2019-12-16","2019-12-16","2019-12-16","2019-12-17","2019-12-17","2019-12-17","2019-12-18","2019-12-18","2019-12-18","2019-12-18","2019-12-18","2019-12-19","2019-12-19","2019-12-19","2019-12-19","2019-12-19","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-2","2019-12-20","2019-12-20","2019-12-20","2019-12-20","2019-12-20","2019-12-20","2019-12-20","2019-12-21","2019-12-21","2019-12-21","2019-12-21","2019-12-21","2019-12-21","2019-12-21","2019-12-21","2019-12-22","2019-12-22","2019-12-22","2019-12-22","2019-12-23","2019-12-23","2019-12-23","2019-12-23","2019-12-23","2019-12-23","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-24","2019-12-25","2019-12-25","2019-12-25","2019-12-25","2019-12-26","2019-12-26","2019-12-26","2019-12-26","2019-12-26","2019-12-26","2019-12-26","2019-12-27","2019-12-27","2019-12-27","2019-12-27","2019-12-27","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-28","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-29","2019-12-3","2019-12-3","2019-12-3","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-30","2019-12-31","2019-12-31","2019-12-31","2019-12-31","2019-12-31","2019-12-31","2019-12-31","2019-12-4","2019-12-4","2019-12-4","2019-12-5","2019-12-5","2019-12-5","2019-12-5","2019-12-5","2019-12-5","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-6","2019-12-7","2019-12-7","2019-12-7","2019-12-7","2019-12-7","2019-12-7","2019-12-7","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-8","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-12-9","2019-2-10","2019-2-10","2019-2-10","2019-2-10","2019-2-10","2019-2-10","2019-2-11","2019-2-11","2019-2-11","2019-2-11","2019-2-11","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-12","2019-2-13","2019-2-13","2019-2-13","2019-2-13","2019-2-13","2019-2-13","2019-2-13","2019-2-14","2019-2-14","2019-2-14","2019-2-15","2019-2-15","2019-2-15","2019-2-15","2019-2-15","2019-2-15","2019-2-16","2019-2-16","2019-2-17","2019-2-17","2019-2-17","2019-2-18","2019-2-18","2019-2-18","2019-2-19","2019-2-19","2019-2-19","2019-2-19","2019-2-19","2019-2-19","2019-2-19","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-2","2019-2-20","2019-2-20","2019-2-20","2019-2-20","2019-2-22","2019-2-22","2019-2-23","2019-2-23","2019-2-23","2019-2-24","2019-2-24","2019-2-24","2019-2-25","2019-2-25","2019-2-25","2019-2-26","2019-2-27","2019-2-27","2019-2-28","2019-2-28","2019-2-3","2019-2-3","2019-2-3","2019-2-4","2019-2-4","2019-2-4","2019-2-4","2019-2-5","2019-2-5","2019-2-5","2019-2-5","2019-2-5","2019-2-5","2019-2-5","2019-2-5","2019-2-6","2019-2-6","2019-2-6","2019-2-6","2019-2-6","2019-2-6","2019-2-6","2019-2-7","2019-2-7","2019-2-7","2019-2-7","2019-2-7","2019-2-7","2019-2-7","2019-2-8","2019-2-8","2019-2-8","2019-2-8","2019-2-8","2019-2-9","2019-2-9","2019-2-9","2019-2-9","2019-2-9","2019-3-1","2019-3-1","2019-3-10","2019-3-10","2019-3-10","2019-3-11","2019-3-11","2019-3-11","2019-3-12","2019-3-12","2019-3-14","2019-3-15","2019-3-15","2019-3-15","2019-3-15","2019-3-15","2019-3-15","2019-3-15","2019-3-15","2019-3-16","2019-3-16","2019-3-16","2019-3-16","2019-3-16","2019-3-16","2019-3-17","2019-3-17","2019-3-17","2019-3-17","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-18","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-19","2019-3-2","2019-3-2","2019-3-2","2019-3-20","2019-3-20","2019-3-20","2019-3-20","2019-3-20","2019-3-21","2019-3-21","2019-3-21","2019-3-22","2019-3-22","2019-3-23","2019-3-23","2019-3-23","2019-3-23","2019-3-23","2019-3-23","2019-3-24","2019-3-24","2019-3-24","2019-3-24","2019-3-25","2019-3-26","2019-3-27","2019-3-29","2019-3-29","2019-3-3","2019-3-3","2019-3-30","2019-3-30","2019-3-30","2019-3-30","2019-3-31","2019-3-31","2019-3-4","2019-3-4","2019-3-5","2019-3-6","2019-3-7","2019-3-7","2019-3-7","2019-3-8","2019-3-8","2019-3-9","2019-3-9","2019-3-9","2019-3-9","2019-3-9","2019-3-9","2019-4-1","2019-4-10","2019-4-11","2019-4-11","2019-4-12","2019-4-13","2019-4-13","2019-4-13","2019-4-13","2019-4-13","2019-4-13","2019-4-14","2019-4-14","2019-4-14","2019-4-15","2019-4-16","2019-4-17","2019-4-17","2019-4-19","2019-4-20","2019-4-21","2019-4-21","2019-4-22","2019-4-23","2019-4-23","2019-4-23","2019-4-23","2019-4-24","2019-4-24","2019-4-26","2019-4-26","2019-4-27","2019-4-27","2019-4-28","2019-4-28","2019-4-28","2019-4-28","2019-4-29","2019-4-3","2019-4-3","2019-4-3","2019-4-3","2019-4-4","2019-4-4","2019-4-4","2019-4-4","2019-4-4","2019-4-5","2019-4-7","2019-4-7","2019-4-8","2019-4-8","2019-4-9","2019-4-9","2019-4-9","2019-4-9","2019-5-10","2019-5-11","2019-5-11","2019-5-16","2019-5-17","2019-5-18","2019-5-18","2019-5-20","2019-5-20","2019-5-21","2019-5-23","2019-5-25","2019-5-26","2019-5-27","2019-5-28","2019-5-29","2019-5-29","2019-5-29","2019-5-29","2019-5-29","2019-5-3","2019-5-3","2019-5-30","2019-5-4","2019-5-6","2019-5-7","2019-5-7","2019-5-7","2019-5-7","2019-5-9","2019-5-9","2019-5-9","2019-5-9","2019-5-9","2019-5-9","2019-6-1","2019-6-11","2019-6-12","2019-6-14","2019-6-17","2019-6-18","2019-6-19","2019-6-21","2019-6-21","2019-6-23","2019-6-23","2019-6-30","2019-6-30","2019-6-4","2019-6-4","2019-6-4","2019-6-5","2019-6-6","2019-6-7","2019-6-8","2019-6-8","2019-6-9","2019-7-1","2019-7-11","2019-7-12","2019-7-16","2019-7-17","2019-7-17","2019-7-17","2019-7-17","2019-7-18","2019-7-19","2019-7-20","2019-7-21","2019-7-21","2019-7-21","2019-7-21","2019-7-23","2019-7-23","2019-7-24","2019-7-25","2019-7-28","2019-7-28","2019-7-28","2019-7-28","2019-7-28","2019-7-30","2019-7-31","2019-7-31","2019-7-8","2019-7-8","2019-8-10","2019-8-11","2019-8-11","2019-8-14","2019-8-14","2019-8-14","2019-8-15","2019-8-15","2019-8-15","2019-8-16","2019-8-16","2019-8-18","2019-8-19","2019-8-19","2019-8-2","2019-8-2","2019-8-2","2019-8-20","2019-8-20","2019-8-20","2019-8-21","2019-8-22","2019-8-23","2019-8-23","2019-8-24","2019-8-25","2019-8-25","2019-8-25","2019-8-25","2019-8-26","2019-8-26","2019-8-27","2019-8-27","2019-8-28","2019-8-29","2019-8-29","2019-8-29","2019-8-29","2019-8-3","2019-8-3","2019-8-30","2019-8-30","2019-8-30","2019-8-31","2019-8-31","2019-8-5","2019-8-6","2019-8-7","2019-8-7","2019-8-8","2019-8-9","2019-9-11","2019-9-11","2019-9-12","2019-9-12","2019-9-13","2019-9-14","2019-9-14","2019-9-15","2019-9-15","2019-9-17","2019-9-17","2019-9-19","2019-9-2","2019-9-2","2019-9-20","2019-9-22","2019-9-23","2019-9-23","2019-9-23","2019-9-23","2019-9-23","2019-9-24","2019-9-25","2019-9-26","2019-9-26","2019-9-27","2019-9-27","2019-9-28","2019-9-28","2019-9-29","2019-9-29","2019-9-29","2019-9-30","2019-9-30","2019-9-4","2019-9-4","2019-9-5","2019-9-6","2019-9-7","2019-9-7","2019-9-8","2019-9-8","2019-9-8","2019-9-8","2019-9-9","2019-9-9","2020-1-1","2020-1-1","2020-1-1","2020-1-1","2020-1-1","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-10","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-11","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-12","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-13","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-14","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-15","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-16","2020-1-17","2020-1-17","2020-1-17","2020-1-17","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-18","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-19","2020-1-2","2020-1-2","2020-1-2","2020-1-20","2020-1-20","2020-1-20","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-21","2020-1-22","2020-1-22","2020-1-22","2020-1-22","2020-1-22","2020-1-22","2020-1-22","2020-1-23","2020-1-23","2020-1-23","2020-1-23","2020-1-23","2020-1-23","2020-1-23","2020-1-24","2020-1-25","2020-1-25","2020-1-26","2020-1-26","2020-1-26","2020-1-27","2020-1-27","2020-1-27","2020-1-27","2020-1-28","2020-1-28","2020-1-28","2020-1-28","2020-1-29","2020-1-29","2020-1-29","2020-1-29","2020-1-29","2020-1-29","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-3","2020-1-30","2020-1-30","2020-1-30","2020-1-30","2020-1-30","2020-1-31","2020-1-31","2020-1-31","2020-1-4","2020-1-4","2020-1-4","2020-1-4","2020-1-4","2020-1-4","2020-1-4","2020-1-4","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-5","2020-1-6","2020-1-6","2020-1-6","2020-1-6","2020-1-7","2020-1-7","2020-1-7","2020-1-7","2020-1-8","2020-1-8","2020-1-8","2020-1-8","2020-1-8","2020-1-8","2020-1-9","2020-1-9","2020-1-9","2020-1-9","2020-1-9","2020-2-1","2020-2-1","2020-2-1","2020-2-1","2020-2-10","2020-2-10","2020-2-11","2020-2-11","2020-2-11","2020-2-11","2020-2-11","2020-2-11","2020-2-11","2020-2-12","2020-2-12","2020-2-12","2020-2-12","2020-2-12","2020-2-13","2020-2-13","2020-2-13","2020-2-13","2020-2-13","2020-2-13","2020-2-14","2020-2-14","2020-2-14","2020-2-14","2020-2-14","2020-2-14","2020-2-15","2020-2-15","2020-2-15","2020-2-15","2020-2-15","2020-2-15","2020-2-16","2020-2-16","2020-2-16","2020-2-16","2020-2-16","2020-2-16","2020-2-16","2020-2-17","2020-2-17","2020-2-18","2020-2-18","2020-2-18","2020-2-18","2020-2-18","2020-2-18","2020-2-18","2020-2-19","2020-2-19","2020-2-19","2020-2-19","2020-2-19","2020-2-2","2020-2-2","2020-2-2","2020-2-2","2020-2-2","2020-2-20","2020-2-20","2020-2-20","2020-2-20","2020-2-20","2020-2-20","2020-2-21","2020-2-21","2020-2-21","2020-2-22","2020-2-22","2020-2-22","2020-2-22","2020-2-23","2020-2-24","2020-2-24","2020-2-25","2020-2-25","2020-2-25","2020-2-25","2020-2-26","2020-2-27","2020-2-28","2020-2-28","2020-2-28","2020-2-28","2020-2-28","2020-2-29","2020-2-29","2020-2-29","2020-2-29","2020-2-29","2020-2-29","2020-2-29","2020-2-3","2020-2-3","2020-2-3","2020-2-3","2020-2-3","2020-2-3","2020-2-3","2020-2-4","2020-2-4","2020-2-4","2020-2-4","2020-2-4","2020-2-5","2020-2-5","2020-2-5","2020-2-5","2020-2-6","2020-2-6","2020-2-6","2020-2-6","2020-2-6","2020-2-7","2020-2-7","2020-2-7","2020-2-7","2020-2-7","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-8","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-2-9","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-1","2020-3-10","2020-3-10","2020-3-10","2020-3-10","2020-3-11","2020-3-11","2020-3-11","2020-3-11","2020-3-12","2020-3-12","2020-3-12","2020-3-12","2020-3-12","2020-3-13","2020-3-13","2020-3-13","2020-3-13","2020-3-13","2020-3-14","2020-3-14","2020-3-14","2020-3-14","2020-3-15","2020-3-15","2020-3-15","2020-3-15","2020-3-15","2020-3-15","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-16","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-17","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-18","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-19","2020-3-2","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-20","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-21","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-22","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-23","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-24","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-25","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-26","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-27","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-28","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-29","2020-3-3","2020-3-3","2020-3-3","2020-3-3","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-30","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-31","2020-3-4","2020-3-4","2020-3-4","2020-3-4","2020-3-4","2020-3-5","2020-3-5","2020-3-6","2020-3-6","2020-3-7","2020-3-7","2020-3-8","2020-3-8","2020-3-8","2020-3-9","2020-3-9","2020-3-9","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-1","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-10","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-11","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-12","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-13","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-14","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-15","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-16","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-17","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-18","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-19","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-2","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-20","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-21","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-22","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-23","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-24","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-25","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-26","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-27","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-28","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-29","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-3","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-30","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-4","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-5","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-6","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-7","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-8","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-4-9","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-1","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-2","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-3","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-4","2020-5-5","2020-5-5","2020-5-5","2020-5-5","2020-5-5","2020-5-5","2020-5-5","2020-5-5","2020-5-6","2020-5-6","2020-5-6","2020-5-6"];
-    var list=[];
-    var current = null;
-            var cnt = 0;
-            for (var i = 0; i < joins.length; i++) {
-              if (joins[i] != current) {
-                if (cnt > 0) {
-                  list.push(current + ": " + cnt);
-                }
-                current = joins[i];
-                cnt = 1;
-              } else {
-                cnt++;
-              }
-            }
-            if (cnt > 0) {
-              list.push(current + ": " + cnt);
-            }
-    fs.writeFile("Output.txt", JSON.stringify(list), function(err) {
-            console.error;
-          });
-    message.reply("done");
-  }
-};
