+++ commands/time.js
@@ -1,6 +1,6 @@
 module.exports = {
   name: "time",
-  description: "What's the time Mr Wolf?",
+  description: "What's the time?",
   usage: "",
   execute(message, args) {
     var time = new Date();
