+++ handler.js
@@ -1,0 +1,27 @@
+const fs = require("fs");
+const Discord = require("discord.js");
+const client = new Discord.Client({partials: ["MESSAGE", "CHANNEL", "REACTION"]});
+const prefix = process.env.PREFIX;
+client.commands = new Discord.Collection();
+
+const commandFiles = fs.readdirSync("./commands").filter(file => file.endsWith(".js"));
+for (const file of commandFiles) {
+  const command = require(`./commands/${file}`);
+  client.commands.set(command.name, command);
+}
+module.exports = function(message) {
+  if (message.author.bot) return;
+  if (!message.content.startsWith(prefix)) return;
+  const args = message.content.slice(prefix.length).split(/ +/);
+  const commandName = args.shift().toLowerCase();
+  const command = client.commands.get(commandName) || client.commands.find( cmd => cmd.aliases && cmd.aliases.includes(commandName) );
+  if (!command) return;
+  if (command.args && !args.length) {
+    let reply = `You didn't provide any arguments, ${message.author}!`;
+    return message.channel.send(reply);
+  }
+  try { command.execute(message, args);
+  } catch (error) {
+    console.error(error);
+    message.reply("there was an error trying to execute that command!");
+  }};
