--- rss/501890309039325224.json
@@ -1,22 +1,0 @@
-{
-  "version": "5.0.3",
-  "name": "Zwift Insider",
-  "id": "501890309039325224",
-  "sources": {
-    "-2147274112zwiftinsidercom-76247": {
-      "title": "Zwift Insider",
-      "link": "https://zwiftinsider.com/feed/",
-      "channel": "644071718314311680",
-      "addedOn": "2020-03-22T16:40:22.417Z",
-      "lastRefreshRateMin": 30,
-      "subscribers": [
-        {
-          "id": "691291967648301076",
-          "name": "Zwift News",
-          "type": "role"
-        }
-      ],
-      "message": "{subscriptions}{link}"
-    }
-  }
-}
