--- rss/subscribers/5e77761ae35ea52c004dc143.json
@@ -1,8 +1,0 @@
-{
-  "filters": {},
-  "rfilters": {},
-  "feed": "5e776d14706e1a2434ccac4b",
-  "id": "691291967648301076",
-  "type": "role",
-  "_id": "5e77761ae35ea52c004dc143"
-}
