--- watch.json
@@ -1,10 +1,0 @@
-{
-  "install": {
-    "include": ["^package\\.json$", "^\\.env$"]
-  },
-  "restart": {
-    "exclude": ["^public/", "^dist/"],
-    "include": ["\\.js$", "\\.json"]
-  },
-  "throttle": 900000
-}
